Переподключила компоненты к стору
