+++ src/components/app/app.jsx
@@ -1,8 +1,6 @@
 import React from "react";
 import PropTypes from "prop-types";
 import {Switch, Route, BrowserRouter} from "react-router-dom";
-import {connect} from "react-redux";
-import {ActionCreator} from "../../store/action";
 import MainPage from "../main-page/main-page";
 import SignIn from "../sign-in/sign-in";
 import MyList from "../my-list/my-list";
@@ -11,7 +9,7 @@ import AddReview from "../add-review/add-review";
 import Player from "../player/player";
 
 const App = (props) => {
-  const {promoFilm, films, filteredFilms, activeGenre, onGenreClick} = props;
+  const {promoFilm} = props;
 
   return (
     <BrowserRouter>
@@ -21,11 +19,7 @@ const App = (props) => {
           render={({history}) => (
             <MainPage
               promoFilm={promoFilm}
-              films={films}
-              filteredFilms={filteredFilms}
               onFilmCardClick={(id) => history.push(`/films/${id}`)}
-              activeGenre={activeGenre}
-              onGenreClick={onGenreClick}
             />
           )} />
         <Route exact path="/login">
@@ -35,7 +29,6 @@ const App = (props) => {
           path="/mylist"
           render={({history}) => (
             <MyList
-              films={films}
               onFilmCardClick={(id) => history.push(`/films/${id}`)}
             />
           )}
@@ -45,7 +38,6 @@ const App = (props) => {
           render={({history, match}) => (
             <FilmPage
               currentFilmId={match.params.id}
-              films={films}
               onFilmCardClick={(id) => history.push(`/films/${id}`)}
             />
           )}
@@ -56,7 +48,7 @@ const App = (props) => {
           render={({match}) => (
             <AddReview
               currentFilmId={match.params.id}
-              films={films} />
+            />
           )}
         />
         <Route exact path="/player/:id">
@@ -67,26 +59,8 @@ const App = (props) => {
   );
 };
 
-const mapStateToProps = (state) => ({
-  films: state.films,
-  filteredFilms: state.filteredFilms,
-  activeGenre: state.activeGenre,
-});
-
-const mapDispatchToProps = (dispatch) => ({
-  onGenreClick(activeGenre) {
-    dispatch(ActionCreator.changeFilter(activeGenre));
-    dispatch(ActionCreator.getFilmsByGenre(activeGenre));
-  }
-});
-
 App.propTypes = {
-  films: PropTypes.array.isRequired,
-  filteredFilms: PropTypes.array.isRequired,
   promoFilm: PropTypes.object.isRequired,
-  activeGenre: PropTypes.string.isRequired,
-  onGenreClick: PropTypes.func.isRequired
 };
 
-export {App};
-export default connect(mapStateToProps, mapDispatchToProps)(App);
+export default App;
